requirements.txt
Django==2.0.5
gdata==2.0.18
isodate==0.6.0
markdown2==2.3.5
pip-tools==2.0.2
pytz==2018.4
smartypants==2.0.1
python-magic==0.4.15
